+++ CHANGELOG.md
@@ -81,3 +81,5 @@
 - Añadido clase TexturaSolidaFactory para reutilizar la creacion de fondos solidos y lisos
 - Refactorizacion de VentanaSeleccionRival para evitar duplicidad de codigo
 - Refactorizacion de la clase Jugador, creando la clase RepositorioJugador para manejo de carga y guardado de datos
+- Refactorizacion Basicos, Palanca y Semaforo con ElementoHUD por el campo OrtographicCamera. 
+-